scrapping/debates-hoc-uk/requirements.txt
attrs==21.2.0
beautifulsoup4==4.9.3
boto3==1.18.29
botocore==1.21.29
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
fuzzywuzzy==0.18.0
html5lib==1.1
idna==3.2
jmespath==0.10.0
jsonschema==3.2.0
pynamodb==5.1.0
pyrsistent==0.18.0
python-dateutil==2.8.2
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.6
webencodings==0.5.1
xmltodict==0.12.0
python-Levenshtein==0.12.2
parsel~=1.5.2